chore: addressing PR feedback
